refactor RequestDecoder protocol
