+++ lib/mongoose_push_web/schemas/request/send_notification/deep/silent_notification.ex
@@ -15,33 +15,15 @@ defmodule MongoosePushWeb.Schemas.Request.SendNotification.Deep.SilentNotificati
 
   defimpl RequestDecoder,
     for: MongoosePushWeb.Schemas.Request.SendNotification.Deep.SilentNotification do
+    alias MongoosePushWeb.Protocols.RequestDecoderHelper
+
+    @spec decode(%Deep.SilentNotification{}) :: MongoosePush.request()
     def decode(schema) do
       %{
         service: String.to_atom(schema.service),
         data: schema.data
       }
-      |> add_optional_fields(schema)
+      |> RequestDecoderHelper.add_optional_fields(schema)
     end
-
-    defp add_optional_fields(push_request, schema) do
-      opt_keys = [:mutable_content, :mode, :priority, :tags, :time_to_live, :topic]
-
-      Enum.reduce(opt_keys, push_request, fn x, acc ->
-        case Map.get(schema, x) do
-          nil -> acc
-          val -> Map.put(acc, x, maybe_parse_to_atom(x, val))
-        end
-      end)
-    end
-
-    defp maybe_parse_to_atom(:mode, val), do: parse_mode(val)
-    defp maybe_parse_to_atom(:priority, val), do: parse_priority(val)
-    defp maybe_parse_to_atom(_key, val), do: val
-
-    defp parse_mode("prod"), do: :prod
-    defp parse_mode("dev"), do: :dev
-
-    defp parse_priority("normal"), do: :normal
-    defp parse_priority("high"), do: :high
   end
 end
